requirements.txt
fastapi
uvicorn
python-multipart
torch
torchaudio
pyannote.audio
transformers